final_project/code/testing/misc/requirements_common.txt
pytest~=6.2.3
pytest-xdist~=2.2.1
allure-pytest~=2.8.40
